+++ core/descriptors.jvm/src/org/jetbrains/kotlin/load/java/JvmAbi.java
@@ -8,6 +8,7 @@ package org.jetbrains.kotlin.load.java;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.kotlin.builtins.CompanionObjectMapping;
 import org.jetbrains.kotlin.descriptors.*;
+import org.jetbrains.kotlin.metadata.jvm.deserialization.JvmProtoBufUtil;
 import org.jetbrains.kotlin.name.ClassId;
 import org.jetbrains.kotlin.name.FqName;
 import org.jetbrains.kotlin.name.Name;
@@ -43,7 +44,7 @@ public final class JvmAbi {
     public static final String INSTANCE_FIELD = "INSTANCE";
     public static final String HIDDEN_INSTANCE_FIELD = "$$" + INSTANCE_FIELD;
 
-    public static final String DEFAULT_MODULE_NAME = "main";
+    public static final String DEFAULT_MODULE_NAME = JvmProtoBufUtil.DEFAULT_MODULE_NAME;
     public static final ClassId REFLECTION_FACTORY_IMPL = ClassId.topLevel(new FqName("kotlin.reflect.jvm.internal.ReflectionFactoryImpl"));
 
     public static final String LOCAL_VARIABLE_NAME_PREFIX_INLINE_ARGUMENT = "$i$a$";