+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.0.5
 
 - [`KT-25371`](https://youtrack.jetbrains.com/issue/KT-25371) Support unsigned integers in kotlinx-metadata-jvm
+- [`KT-28682`](https://youtrack.jetbrains.com/issue/KT-28682) Wrong character replacement in ClassName.jvmInternalName of kotlinx-metadata-jvm
 
 ## 0.0.4